cleanup traits
