+++ src/Transformer/HasIncludesTrait.php
@@ -21,6 +21,13 @@ trait HasIncludesTrait
     protected array $defaultIncludes = [];
 
     /**
+     * IncludeMethod name must be built once.
+     *
+     * @var array<string, string>
+     */
+    private array $includeMethodCache = [];
+
+    /**
      * Getter for availableIncludes.
      */
     public function getAvailableIncludes(): array
@@ -163,5 +170,29 @@ trait HasIncludesTrait
         $this->defaultIncludes = $defaultIncludes;
 
         return $this;
+    }
+
+    private function buildMethodName(string $includeName): string
+    {
+        if (!isset($this->includeMethodCache[$includeName])) {
+            // Check if the method name actually exists
+            $methodName = 'include' . str_replace(
+                    ' ',
+                    '',
+                    ucwords(str_replace(
+                        '_',
+                        ' ',
+                        str_replace(
+                            '-',
+                            ' ',
+                            $includeName
+                        )
+                    ))
+                );
+
+            $this->includeMethodCache[$includeName] = $methodName;
+        }
+
+        return $this->includeMethodCache[$includeName];
     }
 }
